requirements.txt
streamlit==1.11.0
mysql-connector-python==8.0.29
pandas==1.4.3
plotly==5.6.0 
pickle
datetime
optparse
turtle
unicodedata
